move crypto logic from storage to azure_cert_manager, add log info and refactor the code
