+++ deploy/mysql_db_schema.html
@@ -307,11 +307,6 @@ hr {
 	</tr>
 	<tr>
 		<td class='spacer'></td>
-		<td class='subname'>azure_pem_url</td>
-		<td class='type'>VARCHAR(200)</td>
-	</tr>
-	<tr>
-		<td class='spacer'></td>
 		<td class='subname'>subscription_id</td>
 		<td class='type'>VARCHAR(100)</td>
 	</tr>
