+++ deploy/mysql_db_schema.json
@@ -204,17 +204,6 @@
         {
           "dataType": "VARCHAR",
           "decimalDigits": 0,
-          "width": "(200)",
-          "name": "azure_pem_url",
-          "nullable": true,
-          "remarks": "",
-          "autoIncremented": false,
-          "databaseSpecificType": "VARCHAR",
-          "size": 200
-        },
-        {
-          "dataType": "VARCHAR",
-          "decimalDigits": 0,
           "width": "(100)",
           "name": "subscription_id",
           "nullable": true,
