+++ open-hackathon-server/src/hackathon/__init__.py
@@ -153,7 +153,7 @@ def init_components():
     from hackathon.database.db_adapters import SQLAlchemyAdapter
     from hackathon.user import UserManager, UserProfileManager
     from hackathon.hack import HackathonManager, AdminManager, TeamManager, DockerHostManager, \
-        AzureCertManager, RegisterManager, HackathonTemplateManager
+        AzureCertManager, RegisterManager, HackathonTemplateManager, Cryptor
     from hackathon.template import TemplateLibrary
     from hackathon.remote.guacamole import GuacamoleInfo
     from hackathon.expr.expr_mgr import ExprManager
@@ -181,6 +181,7 @@ def init_components():
     factory.provide("hackathon_manager", HackathonManager)
     factory.provide("register_manager", RegisterManager)
     factory.provide("azure_cert_manager", AzureCertManager)
+    factory.provide("cryptor", Cryptor)
     factory.provide("docker_host_manager", DockerHostManager)
     factory.provide("hackathon_template_manager", HackathonTemplateManager)
     factory.provide("template_library", TemplateLibrary)
@@ -258,8 +259,6 @@ def init_hackathon_storage():
         from hackathon.hazure import BlobServiceAdapter
         factory.provide("azure_blob_service", BlobServiceAdapter)
         factory.provide("storage", AzureStorage)
-        factory.provide("cryptor", get_class("hackathon.storage.cryptor.Cryptor"))
-
     else:
         factory.provide("storage", LocalStorage)
