+++ open-hackathon-server/src/hackathon/azureformation/service.py
@@ -64,12 +64,7 @@ class Service(ServiceManagementService, Component):
     def __init__(self, azure_key_id):
         self.azure_key_id = azure_key_id
         azure_key = self.db.get_object(AzureKey, self.azure_key_id)
-
-        if not isfile(azure_key.pem_url):
-            cryptor = RequiredFeature("cryptor")
-            cryptor.recover_local_file(azure_key.azure_pem_url, azure_key.pem_url)
-
-        super(Service, self).__init__(azure_key.subscription_id, azure_key.pem_url, azure_key.management_host)
+        super(Service, self).__init__(azure_key.subscription_id, azure_key.get_local_pem_url(), azure_key.management_host)
 
     # ---------------------------------------- subscription ---------------------------------------- #
