+++ open-hackathon-server/src/hackathon/hack/host_server_manager.py
@@ -452,11 +452,8 @@ class DockerHostManager(Component):
             self.log.error('Found no azure key with Hackathon:%d' % hackathon_id)
             return None
 
-        if not isfile(hackathon_azure_key.azure_key.pem_url):
-            cryptor = RequiredFeature("cryptor")
-            cryptor.recover_local_file(hackathon_azure_key.azure_key.azure_pem_url, hackathon_azure_key.azure_key.pem_url)
         sms = ServiceManagementService(hackathon_azure_key.azure_key.subscription_id,
-                                       hackathon_azure_key.azure_key.pem_url,
+                                       hackathon_azure_key.azure_key.get_local_pem_url(),
                                        host=hackathon_azure_key.azure_key.management_host)
         return sms
